+++ CPUVersion/run/main.cpp
@@ -35,7 +35,7 @@ int main() {
                 0.5   // wyh
   );                  // thr (对于WFR不需要)
 
-  std::vector<double> result;
+  std::vector<float> result;
 
   std::cout << "Please input any key to continue" << std::endl;
   std::cin.get();
